fix: address well-architected review feedback
